Refactor scan logic with 2-byte constraint, integrate student metadata, and enhance UI feedback
- Implement duplicate check based on student number's last 5 digits (ushort).
- Add StudentService to resolve student name and code from CSV.
- Add flashing animations for input errors and highlighting for new records.
- Ensure binary storage uses 2-byte UInt16 (ushort) format as requested.
- Improve notification system with color-coded status bar and re-trigger logic.
- Implement INotifyPropertyChanged in ScanRecord for UI responsiveness.
- Use safe parsing (ushort.TryParse) for manual input handling.

+++ Models/ScanRecord.cs
@@ -10,7 +10,7 @@ namespace Tenko.Native.Models
         public string Id { get; set; } = string.Empty;
         public DateTime Timestamp { get; set; }
         public string Barcode { get; set; } = string.Empty;
-        public int Last5 { get; set; }
+        public ushort Last5 { get; set; }
         public string Location { get; set; } = string.Empty;
 
         // Metadata resolved from student data
